requirements.txt
django==5.2.11
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
channels==4.3.2
paho-mqtt==2.1.0
